requirements.txt
dj-database-url==0.3.0
Django==1.8.3
psycopg2==2.6.1
python-decouple==2.3
pytz==2015.4
Unipath==1.1
python-firebase==1.2
requests==2.7.0
beautifulsoup4==4.4.0